agentic_conversation_agent/rootfs/app/requirements.txt
aiohttp==3.10.10
pyyaml==6.0.2
jsonschema==4.23.0
openai==1.57.3
google-generativeai==0.8.3
